Added level Select Screen and menu input handler
The input handler helps deal with insuring that pressing a button (like
up or enter) will only count as one press. Seems to be buggy on hitting
escape which is why the pause menu doesn't work right now. If somebody
could find or make an actual background image for the level select
screen, that would be great.