Fixed Pause Screen and added background
EventController and all of its events seem to be functional. I also
added some lousy backgrounds to spruce it up a bit. If anybody has
better images, feel free to replace them. Also, the water level that was
added doesn't work, which might mean that there are tiles on the ground
layer that don't have the three required types: type, left, and right.