Fixes from Grant's commit (read below!)
The way the tile_engine works requires each map to be done a certain
way. Please look at the readme.txt inside levels and use the
tiled_5_3_grassy_plains.tmx as an example for making your map correctly.
Also, I had to comment out the change to game_events for the game to
actually work since I had no idea what it does and how to conciliate it
with my changes. Feel free to change it back (with comments would be
appreciated) if it's needed.